requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cloudinary==1.29.0
Django==4.0.3
django-phonenumber-field==6.1.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
phonenumbers==8.12.45
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
